services/api/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
kafka-python>=2.0.2,<3.0.0
prometheus-fastapi-instrumentator>=6.1.0,<7.0.0
httpx>=0.27.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
redis[hiredis]>=5.0.0,<6.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
